fix: name component
